Make intended internal classes truly internal.
`InternalObjectExtensions` and `InternalTypeExtensions` were only public for the sake of unit testing.
Since #1434 FluentAssertions.Specs is also strong named, so these classes can now be truly internal.
