+++ Src/FluentAssertions/Collections/CollectionAssertions.cs
@@ -1626,7 +1626,7 @@ namespace FluentAssertions.Collections
         /// </param>
         public AndConstraint<TAssertions> StartWith(object element, string because = "", params object[] becauseArgs)
         {
-            AssertCollectionStartsWith(Subject?.Cast<object>(), new[] { element }, InternalObjectExtensions.IsSameOrEqualTo, because, becauseArgs);
+            AssertCollectionStartsWith(Subject?.Cast<object>(), new[] { element }, ObjectExtensions.IsSameOrEqualTo, because, becauseArgs);
             return new AndConstraint<TAssertions>((TAssertions)this);
         }
 
@@ -1680,7 +1680,7 @@ namespace FluentAssertions.Collections
         /// </param>
         public AndConstraint<TAssertions> EndWith(object element, string because = "", params object[] becauseArgs)
         {
-            AssertCollectionEndsWith(Subject?.Cast<object>(), new[] { element }, InternalObjectExtensions.IsSameOrEqualTo, because, becauseArgs);
+            AssertCollectionEndsWith(Subject?.Cast<object>(), new[] { element }, ObjectExtensions.IsSameOrEqualTo, because, becauseArgs);
             return new AndConstraint<TAssertions>((TAssertions)this);
         }
