+++ Src/FluentAssertions/Common/ObjectExtensions.cs
@@ -3,8 +3,7 @@ using System.Globalization;
 
 namespace FluentAssertions.Common
 {
-    // NOTE: These extension methods are covered by tests separately and therefore need to be public.
-    public static class InternalObjectExtensions
+    internal static class ObjectExtensions
     {
         public static bool IsSameOrEqualTo(this object actual, object expected)
         {
